task o

+++ 0x12-more_singly_linked_lists/0-print_listint.c
@@ -1,3 +1,6 @@
+#include <stdlib.h>
+#include <string.h>
+#include <stdio.h>
 #include "lists.h"
 /**
  * print_listint - Print
@@ -7,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	unsigned int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
